Refactor default rules into a preset
