+++ src/lint/presets.ts
@@ -1,0 +1,14 @@
+import {
+  RulesPreset,
+  subjectMaxLength,
+  subjectMinLength,
+  subjectCapitalizeFirst,
+  subjectNoEndPeriod,
+} from "lint/rules";
+
+export const defaultPreset: RulesPreset = [
+  subjectMaxLength,
+  subjectMinLength,
+  subjectCapitalizeFirst,
+  subjectNoEndPeriod,
+];
