+++ src/lint/rules/subject-min-length.ts
@@ -1,18 +1,18 @@
-import { RuleCheckResults } from "types/rules";
+import { RuleCheckResults } from "./";
 import { Commit } from "@commitlint/parse";
 
-const maxLength = 50;
+const minLength = 0;
 
 export default {
-  name: "subject-max-length",
+  name: "subject-min-length",
   score: 5,
   check: (commit: Commit): RuleCheckResults => {
     const actualLength = commit.header.length;
-    return actualLength <= maxLength
+    return actualLength >= minLength
       ? null
       : {
           actualLength,
-          maxLength,
+          minLength,
         };
   },
 };
